backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
django
django-cors-headers==3.8.0
django-rest-framework
django-silk
executing==2.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langsmith==0.1.147
marshmallow==3.23.1
matplotlib-inline==0.1.7
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
numpy==2.1.3
openai==1.57.0
orjson==3.10.12
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
tenacity==9.0.0
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.18.3
